* upstream/master: (42 commits)
  Avoid reference cycle
  Improve TemplateTypeScope interface
  New Musement link
  Bump cache
  ConstantArrayTypeBuilder solve nextAutoIndex by itself
  This is more future-proof
  Fix CS
  GenericParametersAcceptorResolver - how do constant types behave?
  Drop ParametersAcceptorSelector::selectArguments
  Increase memory limit
  Use ParametersAcceptorSelector::selectArguments() to find internal function types
  ParametersAcceptorSelector: Resolve template types
  Add GenericParametersAcceptorResolver
  Reflection: support parameter default values
  Open 0.12-dev
  Updated zip_entry_read return type
  Ignore files using a stream wrapper. (#2184)
  exec(): use PHP_BINARY instead of env php (#2185)
  Fixed array non-emptyness after array_pop/push/shift/unshift
  is_callable() with an array with two items results in method_exists()
  ...

+++ src/Type/UnionTypeHelper.php
@@ -67,7 +67,7 @@ class UnionTypeHelper
 	 */
 	public static function sortTypes(array $types): array
 	{
-		usort($types, static function (Type $a, Type $b): float {
+		usort($types, static function (Type $a, Type $b): int {
 			if ($a instanceof NullType) {
 				return 1;
 			} elseif ($b instanceof NullType) {
@@ -108,7 +108,7 @@ class UnionTypeHelper
 					|| $b instanceof ConstantFloatType
 				)
 			) {
-				return $a->getValue() - $b->getValue();
+				return (int) ($a->getValue() - $b->getValue());
 			}
 
 			if ($a instanceof ConstantStringType && $b instanceof ConstantStringType) {
